Code cleanup_02

+++ src/main/java/keystrokesmod/utils/RenderUtils.java
@@ -97,7 +97,7 @@ public class RenderUtils {
     }
 
     public static void drawBorderedRect(float f, float f1, float f2, float f3, float f4, int i, int j) {
-        drawRect((double) f, (double) f1, (double) f2, (double) f3, j);
+        drawRect(f, f1, f2, f3, j);
         float f5 = (float) (i >> 24 & 255) / 255.0F;
         float f6 = (float) (i >> 16 & 255) / 255.0F;
         float f7 = (float) (i >> 8 & 255) / 255.0F;
@@ -111,14 +111,14 @@ public class RenderUtils {
         GL11.glColor4f(f6, f7, f8, f5);
         GL11.glLineWidth(f4);
         GL11.glBegin(1);
-        GL11.glVertex2d((double) f, (double) f1);
-        GL11.glVertex2d((double) f, (double) f3);
-        GL11.glVertex2d((double) f2, (double) f3);
-        GL11.glVertex2d((double) f2, (double) f1);
-        GL11.glVertex2d((double) f, (double) f1);
-        GL11.glVertex2d((double) f2, (double) f1);
-        GL11.glVertex2d((double) f, (double) f3);
-        GL11.glVertex2d((double) f2, (double) f3);
+        GL11.glVertex2d(f, f1);
+        GL11.glVertex2d(f, f3);
+        GL11.glVertex2d(f2, f3);
+        GL11.glVertex2d(f2, f1);
+        GL11.glVertex2d(f, f1);
+        GL11.glVertex2d(f2, f1);
+        GL11.glVertex2d(f, f3);
+        GL11.glVertex2d(f2, f3);
         GL11.glEnd();
         GL11.glPopMatrix();
         GL11.glEnable(GL11.GL_TEXTURE_2D);
